testing animation smoothing
Still exploring this but wanted to start sharing. Based on discussion here: https://2dimensions.slack.com/archives/C078KG6KAJY/p1729226510912809

Diffs=
76f79a22e2 testing animation smoothing (#8363)

+++ include/rive/artboard.hpp
@@ -140,10 +140,8 @@ public:
     bool syncStyleChanges();
     bool canHaveOverrides() override { return true; }
 
-    bool advance(double elapsedSeconds,
-                 bool nested = true,
-                 bool animate = true);
-    bool advanceInternal(double elapsedSeconds,
+    bool advance(float elapsedSeconds, bool nested = true, bool animate = true);
+    bool advanceInternal(float elapsedSeconds,
                          bool isRoot,
                          bool nested = true,
                          bool animate = true);
